[Refactor] LIVD-456 - 캘린더 WebView Helper 분리와 높이·파서 정리
Coordinator 책임을 LoadSession·HeightMeasurer로 나누고, inject 직전 높이 fallback 리셋과 메시지 body 파서 공통화를 적용한다.
