+++ docs/archives/LIVD-456-home-calendar-webview.md
@@ -1,7 +1,7 @@
 # LIVD-456 홈 캘린더 WebView 연동
 
 > 통합 아카이브. 기존 분리 계획  
-> (`home-calendar-webview` · `calendar-month-changed` · `calendar-month-restore` · `day-event-id-time`)를 Phase로 합쳤다.  
+> (`home-calendar-webview` · `calendar-month-changed` · `calendar-month-restore` · `day-event-id-time` · `calendar-webview-split-height`)를 Phase로 합쳤다.  
 > 트러블슈팅: `docs/archives/LIVD-456-home-calendar-webview-troubleshooting.md`  
 > 브랜치: `feat/LIVD-456-home-calendar-webview`
 
@@ -10,6 +10,7 @@
 - 월 그리드는 WebView, Store `calendarMonth`는 브릿지로 주입해야 한다.
 - 웹 계약: iOS→Web `setCalendarData(객체)`, Web→iOS `calendarDateSelected` / `calendarMonthChanged`.
 - 이후 상세 복귀 시 월 리셋, 동일 concertID·다른 time 행 누락을 같은 이슈에서 수정했다.
+- 머지 전 개선: Coordinator 비대화·높이 고착·메시지 body 파서 중복을 정리했다.
 
 ## 목표 (전체)
 - `CALENDAR_WEB_URL`로 WebView 로드 (없/실패 → `about:blank`, 앱 비종료).
@@ -17,6 +18,7 @@
 - 상세·탭 복귀 시 선택 월 유지 + soft refresh (관심 변경 반영).
 - 일자 모달에서 동일 concert·다른 시각 행을 모두 표시.
 - 일자 일정 → 콘서트 상세 이동. DEBUG 「일정 모달」 버튼 제거.
+- WebView Helper 분리·inject 직전 높이 fallback 리셋·메시지 body 파서 공통화 (기존 동작 유지).
 
 ---
 
@@ -108,26 +110,59 @@
 
 ---
 
+## Phase 5 — WebView 분리 · 높이 리셋 · 파서 공통화
+
+### 목표
+- Representable API·브릿지·Store 동작은 유지한 채 Coordinator를 Helper로 분리한다.
+- 새 페이로드 inject 직전 `contentHeight`를 fallback(700)으로 리셋한다.
+- WK 메시지 body → dictionary / Int 변환을 공통 Helper로 모은다.
+
+### 작업 항목
+- [x] 보호 테스트 확인 후 리팩터링 시작
+- [x] `CalendarWebScriptMessageBodyParser` + 테스트 (TDD) — dict / NSDictionary / JSON string, `intValue`
+- [x] DateSelected·MonthChanged 파서 슬림화 (순수 `yyyy-MM-dd` 문자열은 DateSelected에 유지)
+- [x] `CalendarWebContentHeightMeasurer` 분리 (unlock script · maxBottom 측정 · fallbackHeight)
+- [x] `CalendarWebLoadSession` 분리 (load/inject/fail · Gate · inject 직전 높이 리셋)
+- [x] `CalendarWebView` 슬림화 — weak proxy 패턴 유지, dismantle에서 handler 이름별 제거
+- [x] `tuist generate` · HomeFeature 캘린더 테스트 47 passed · `Livith-iOS-Dev` build 성공 · 아카이브 통합
+
+### Phase 5 기술 결정
+| 결정 | 내용 |
+|------|------|
+| 분리 위치 | Helper 파일 (Representable은 배선만) |
+| 높이 리셋 | inject 직전 700, 동일 JSON 스킵 시 리셋 없음 |
+| 웹 height 브릿지 | 비범위 |
+| 파서 공통화 | BodyParser 포함, 도메인 파싱은 각 파서 유지 |
+| TDD | BodyParser red→green, WK 배선·높이 리셋은 예외+보호 테스트 |
+| year 범위 검증 | 비범위 (month 1…12만) |
+
+---
+
 ## 영향 범위 (전체)
 - App: plist, xcconfig 키, `CalendarWebConfig` DI
-- HomeFeature: WebView·Helpers·Store·ContentView·모달 네비게이션
+- HomeFeature: WebView·Helpers(LoadSession·HeightMeasurer·BodyParser 포함)·Store·ContentView·모달 네비게이션
 - Domain: `CalendarEventID` / `CalendarDayEvent`
 - CalendarData: Mapper 테스트 (계약 스키마 변경 없음)
 - 문서: 본 계획 + `LIVD-456-home-calendar-webview-troubleshooting.md`
 
 ## 주의 사항
 - 실 `CALENDAR_WEB_URL` 원문은 문서·PR·채팅에 남기지 않는다.
-- WK 배선은 TDD 예외, 파서·Store·매퍼·Gate·Domain identity는 TDD.
+- WK 배선은 TDD 예외, 파서·Store·매퍼·Gate·Domain identity·BodyParser는 TDD.
 - `setCalendarData`는 실 URL load 완료 전·blank·로드 실패 후 호출하지 않는다.
 - remount 시 monthChanged gate 리셋·재inject 필요.
+- `WeakScriptMessageHandlerProxy` 패턴 유지 + dismantle에서 handler 이름 제거 (혼동 금지).
+- inject 성공 completion에서만 `lastInjected`·Gate 갱신. inject 직전 높이 fallback 리셋.
 
 ## 검증 방법
 - `tuist generate --no-open` (파일 추가 시)
 - Domain / CalendarData / HomeFeature 관련 `xcodebuild test` (simulator: iPhone 17)
 - `Livith-iOS-Dev` build
 - 수동: URL·주입·날짜 탭·월 이동·상세 복귀 월 유지·동일 콘서트 복수 시각 행·콘서트 상세 이동
+- Phase 5 수동: 월 스와이프 시 높이 fallback→재측정, 짧은/긴 달 전환 잘림·공백 없음
 
 ## 비범위 (잔여)
 - 웹 `calendarReady` 등 추가 메시지
 - 네이티브 월 그리드 대체
 - `time == nil`인 동일 concert·type 중복 행 identity
+- 웹 height `postMessage` 브릿지
+- `CalendarWebConfig` 모듈 이동 · year 범위 검증 강화
